backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
requests
jinja2 
asyncpg